% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sbalkus.R
\name{format_sbalkus}
\alias{format_sbalkus}
\title{Salvador Balkus' ggplot formatting function}
\usage{
format_sbalkus(
  palette_ordered = TRUE,
  continuous_axes = TRUE,
  map = FALSE,
  enlarge_legend_items = FALSE,
  colors = "discrete"
)
}
\arguments{
\item{palette_ordered}{Whether to use a color palette ordered by color lightness.}

\item{continuous_axes}{Whether axes are continuous; if so, cap axis at endpoints.}

\item{map}{Whether the ggplot is a map, to remove axes and preserve geographical scaling.}

\item{enlarge_legend_items}{If True, enlarges legend items. Use for scatter plots or other plots with small dots.}

\item{colors}{Whether the variable mapped to color is continuous. Options: "continuous" or "discrete"}
}
\description{
This function modifies the ggplot to have the proper colors, fonts, and other formats.
}
