% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_sbalkus_ordered.R
\docType{data}
\name{palette_sbalkus_ordered}
\alias{palette_sbalkus_ordered}
\title{Salvador Balkus' data visualization theme colors.}
\format{
An object of class \code{character} of length 5.
}
\usage{
palette_sbalkus_ordered
}
\description{
A list of colors used in papers by Salvador Balkus. The colors are ordered by their lightness in grayscale.
}
\keyword{datasets}
