% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_sbalkus_unordered.R
\docType{data}
\name{palette_sbalkus_unordered}
\alias{palette_sbalkus_unordered}
\title{Salvador Balkus' data visualization theme colors.}
\format{
An object of class \code{character} of length 5.
}
\usage{
palette_sbalkus_unordered
}
\description{
A list of colors used in papers by Salvador Balkus. The colors are unordered to plot highly distinguishable colors next to each other.
}
\keyword{datasets}
